Move testing helper classes to new helpers directory
