+++ packages/framework/tests/helpers/BaseMarkdownPageUnitTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Testing\Unit\Pages;
+namespace Hyde\Framework\Testing\helpers;
 
 require_once __DIR__.'/BaseHydePageUnitTest.php';
 
@@ -11,7 +11,7 @@ require_once __DIR__.'/BaseHydePageUnitTest.php';
  *
  * These unit tests ensure all inherited methods are callable, and that they return the expected value.
  *
- * @see \Hyde\Framework\Testing\Unit\Pages\BaseHydePageUnitTest
+ * @see \Hyde\Framework\Testing\helpers\BaseHydePageUnitTest
  *
  * @coversNothing
  */
